feat:gateway 추가
- gateway 추가
- endpoing 수정
