+++ settings.gradle
@@ -6,7 +6,7 @@ include(
         "syncfit-eureka-service",
         "syncfit-gateway-service",
         "syncfit-config-service",
-        "syncfit-outterapi-service",
+        "syncfit-outerapi-service",
         "syncfit-wishlist-service",
         "syncfit-track-service",
         "syncfit-image-service",
