+++ syncfit-outerapi-service/Dockerfile
@@ -1,0 +1,4 @@
+FROM openjdk:17
+ARG JAR_FILE=build/libs/*.jar
+COPY ${JAR_FILE} app.jar
+ENTRYPOINT ["java", "-jar", "/app.jar"]
