+++ syncfit-outerapi-service/gradle/wrapper/gradle-wrapper.properties
