+++ syncfit-outerapi-service/src/main/java/com/syncfitouterapiservice/dto/response/GenreRecommendationResponse.java
@@ -1,0 +1,12 @@
+package com.syncfitouterapiservice.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class GenreRecommendationResponse {
+    private String genre;
+}
