+++ syncfit-outerapi-service/src/main/java/com/syncfitouterapiservice/dto/response/SpotifySearchResponse.java
@@ -1,0 +1,7 @@
+package com.syncfitouterapiservice.dto.response;
+
+public record SpotifySearchResponse(String artistName, String title, String albumName, String imageUrl) {
+    public static SpotifySearchResponse of(String artistName, String title, String albumName, String imageUrl) {
+        return new SpotifySearchResponse(artistName, title, albumName, imageUrl);
+    }
+}
