+++ syncfit-outerapi-service/src/main/java/com/syncfitouterapiservice/dto/response/YouTubeVideoIdResponse.java
@@ -1,0 +1,4 @@
+package com.syncfitouterapiservice.dto.response;
+
+public record YouTubeVideoIdResponse(String videoId) {
+}
