+++ syncfit-outerapi-service/src/test/java/com/syncfitouterapiservice/SyncfitOuterapiServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.syncfitouterapiservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class SyncfitOuterapiServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
