--- syncfit-outterapi-service/src/main/java/com/syncfitoutterapiservice/domain/chat/dto/response/GenreRecommendationResponse.java
@@ -1,12 +1,0 @@
-package com.syncfitoutterapiservice.domain.chat.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class GenreRecommendationResponse {
-    private String genre;
-}
