--- syncfit-outterapi-service/src/main/java/com/syncfitoutterapiservice/domain/spotify/dto/response/SpotifySearchResponse.java
@@ -1,7 +1,0 @@
-package com.syncfitoutterapiservice.domain.spotify.dto.response;
-
-public record SpotifySearchResponse(String artistName, String title, String albumName, String imageUrl) {
-    public static SpotifySearchResponse of(String artistName, String title, String albumName, String imageUrl) {
-        return new SpotifySearchResponse(artistName, title, albumName, imageUrl);
-    }
-}
