--- syncfit-outterapi-service/src/main/java/com/syncfitoutterapiservice/domain/youtube/dto/response/YouTubeVideoIdResponse.java
@@ -1,4 +1,0 @@
-package com.syncfitoutterapiservice.domain.youtube.dto.response;
-
-public record YouTubeVideoIdResponse(String videoId) {
-}
