--- syncfit-outterapi-service/src/main/java/com/syncfitoutterapiservice/infra/config/chat/ChatConfig.java
@@ -1,37 +1,0 @@
-package com.syncfitoutterapiservice.infra.config.chat;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.stereotype.Component;
-import org.springframework.web.client.RestClient;
-import org.springframework.web.client.support.RestClientAdapter;
-import org.springframework.web.service.invoker.HttpServiceProxyFactory;
-
-@Component
-@RequiredArgsConstructor
-public class ChatConfig {
-
-    @Value("${google.gemini-baseurl}")
-    private String baseUrl;
-
-    @Value("${google.api-key}")
-    private String googleApiKey;
-
-    @Bean
-    public RestClient chatRestClient() {
-        return RestClient.builder()
-                .baseUrl(baseUrl)
-                .defaultHeader("x-goog-api-key", googleApiKey)
-                .defaultHeader("Content-Type", "application/json")
-                .defaultHeader("Accept", "application/json")
-                .build();
-    }
-
-    @Bean
-    public ChatInterface chatInterface(RestClient client) {
-        RestClientAdapter adapter = RestClientAdapter.create(client);
-        HttpServiceProxyFactory factory = HttpServiceProxyFactory.builderFor(adapter).build();
-        return factory.createClient(ChatInterface.class);
-    }
-}
