--- syncfit-outterapi-service/src/main/java/com/syncfitoutterapiservice/infra/config/chat/ChatInterface.java
@@ -1,18 +1,0 @@
-package com.syncfitoutterapiservice.infra.config.chat;
-
-import com.syncfitoutterapiservice.domain.chat.dto.request.ChatRequest;
-import com.syncfitoutterapiservice.domain.chat.dto.response.ChatResponse;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.service.annotation.HttpExchange;
-import org.springframework.web.service.annotation.PostExchange;
-
-@HttpExchange("/v1beta/models/")
-public interface ChatInterface {
-
-    @PostExchange("{model}:generateContent")
-    ChatResponse getCompletion(
-            @PathVariable String model,
-            @RequestBody ChatRequest request
-    );
-}
