--- syncfit-outterapi-service/src/main/resources/application.yml
@@ -1,29 +1,0 @@
-server:
-  port: 9999
-
-spring:
-  application:
-    name: outterapi-service
-#  config:
-#    activate:
-#      on-profile: "api"
-
-google:
-  gemini-baseurl: ${GEMINI_BASEURL}
-  api-key: ${GEMINI_API_KEY}
-
-spotify:
-  client-id: ${SPOTIFY_CLIENT_ID}
-  client-secret: ${SPOTIFY_CLIENT_SECRET}
-
-youtube:
-  api-key: ${YOUTUBE_API_KEY}
-
-eureka:
-  instance:
-    instance-id: ${spring.application.name}:${spring.application.instance_id:${random.value}}
-  client:
-    register-with-eureka: true
-    fetch-registry: true
-    service-url:
-      defaultZone: http://localhost:8761/eureka
