--- syncfit-outterapi-service/src/test/java/com/syncfitoutterapiservice/SyncfitOutterapiServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package com.syncfitoutterapiservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SyncfitOutterapiServiceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
